+++ src/sanity/schemaTypes/questionObjects/questionPrompt.ts
@@ -53,4 +53,5 @@ export type QuestionPrompt_t = {
 } | {
 	promptType: "richText",
 	richTextPrompt: PortableTextBlock[]
+	richTextAsPlaintext: string
 }
